Move WlProxy to it's own file
